requirements.txt
python-telegram-bot==22.5
python-dotenv==1.0.0
sqlalchemy
psycopg
tzdata